+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [16.0.3] - 2026-09-17
+
+### Fixed
+
+- Include cancelled deployments in terminal deployment counters and duration metrics using
+  `result="cancelled"`.
+
 ## [16.1.0] - 2026-09-16
 
 ### Dependencies